Fix login page on ios
At least on some iOS devices, the 'Login' button on the /login route was
trying to submit (i.e. actually POST) the form on the page instead of
being intercepted by ember and firing the 'logIn' action.  Added
'onsubmit' and 'onclick' attributes to the form and button
respectively, telling each to do nothing (ember 'action' will pick it up
and run the 'logIn' action).

Because of a different iOS-only browser
quirk, forms' onsubmit attributes are ignored if the form does not
contain an `<input type="submit">`.  Since we want to use a `<button>` for
style reasons, I added a dummy input submit to get the form to read it's
onsubmit attribute properly.